working on setting up the logic to calculate split index which will be handled by the indexing layer but the lower layers need to be callable for this without having knowledge on why

+++ src/page/internal_page.rs
@@ -82,7 +82,7 @@ impl<'page> IndexPageMut<'page> {
 
         // Adjust free_end for special offset
         self.page
-            .set_free_end(PAGE_SIZE as u16 - INDEX_SPECIAL_SIZE)?;
+            .set_free_end(PAGE_SIZE - INDEX_SPECIAL_SIZE as usize)?;
 
         // Set lsn
         self.page.set_lsn(lsn);
